requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
bs4==0.0.1
Pillow==9.5.0
pytest==7.3.1
pytest-rerunfailures==11.1.2
pytest-xdist==3.2.1
selenium==4.8.3
webdriver-manager==3.8.6

